Add in fly-event-processor governor_status postgres repository
